+++ generators/angular-component-list/index.js
@@ -9,6 +9,7 @@ var prompts = require('../../services/prompts');
 var buildContext = require('../../services/build-context');
 var pathNames = require('../../services/path-names');
 var append = require('../../services/append');
+var fs = require('fs');
 
 module.exports = yeoman.Base.extend({
   prompting: function () {
@@ -88,10 +89,12 @@ module.exports = yeoman.Base.extend({
       this.props.angularAppName,
       this.props.angularAppName + '.config.js'
     );
-    this.fs.copy(templatePath, templatePath, {
-      process: function (content) {
-        return append.angularRoute(content, $this.props.objectUrl + 's', $this.props.objectUrl + '-list');
-      }
-    });
+    if (fs.existsSync(templatePath)) {
+      this.fs.copy(templatePath, templatePath, {
+        process: function (content) {
+          return append.angularRoute(content, $this.props.objectUrl + 's', $this.props.objectUrl + '-list');
+        }
+      });
+    }
   }
 });
